requirements.txt
Django==1.9.7
cassandra-driver
pyyaml
